Create new class GetInfo and add 6 methods from main.rb file
